Move expectations to the common pkg and add unit tests
