+++ operators/pkg/controller/elasticsearch/elasticsearch_controller.go
@@ -71,7 +71,7 @@ func newReconciler(mgr manager.Manager, params operator.Parameters) *ReconcileEl
 
 		finalizers:     finalizer.NewHandler(client),
 		dynamicWatches: watches.NewDynamicWatches(),
-		expectations:   driver.NewGenerationExpectations(),
+		expectations:   reconciler.NewExpectations(),
 
 		Parameters: params,
 	}
@@ -170,7 +170,7 @@ type ReconcileElasticsearch struct {
 
 	// expectations help dealing with inconsistencies in our client cache,
 	// by marking resources updates as expected, and skipping some operations if the cache is not up-to-date.
-	expectations *driver.Expectations
+	expectations *reconciler.Expectations
 
 	// iteration is the number of times this controller has run its Reconcile method
 	iteration int64
